Pre: Move query-projectors/errors and aggregates into query-projector-traits
