+++ Cargo.toml
@@ -67,6 +67,9 @@ path = "query-algebrizer-traits"
 [dependencies.mentat_query_projector]
 path = "query-projector"
 
+[dependencies.query_projector_traits]
+path = "query-projector-traits"
+
 [dependencies.mentat_query_pull]
 path = "query-pull"
