+++ query-projector/src/projectors/constant.rs
@@ -20,7 +20,7 @@ use ::{
     rusqlite,
 };
 
-use ::errors::{
+use query_projector_traits::errors::{
     Result,
 };
